main.py: Refactor extract_n_latest_repo_tags a bit
The existing semantics of the code were that for each minor version (e.g
1.2) it creates a list of tags belonging to this version. Then
latest_tags_size tags are returned.
Variable names suggested that this was meant to use major versions, not
minor. This probably made sense before 1.0, but right now it doesn't so
I kept the current approach.

This function has more issues, which I'll fix in next commits.